refactor(extension-common): 取得とメッセージングを移す
credentials と Site Profile のメッセージング定義、フレーム横断のクレデンシャル
取得、Target Integrity 検証器の呼び出しを extension-common へ移す。

UI の props 型 (CredentialsProps / SiteProfileProps) は移さず、それぞれの
コンポーネント側に置いた。検証結果を描画するビューはアプリごとに分かれるため、
共有パッケージには入れない。
