+++ apps/inspector/src/components/CaSelector.tsx
@@ -1,9 +1,12 @@
+import {
+  SupportedVerifiedCa,
+  SupportedVerifiedCas,
+} from "@originator-profile/extension-common";
 import { Image } from "@originator-profile/extension-common/ui";
 import placeholderContentThumbnail from "@originator-profile/extension-common/ui/assets/placeholder-content-thumbnail.png";
 import { Link, useParams } from "react-router";
 import { twMerge } from "tailwind-merge";
 import { buildPublUrl } from "../utils/routes";
-import { SupportedVerifiedCa, SupportedVerifiedCas } from "./credentials";
 
 type Props = {
   filteredCas: SupportedVerifiedCas;
