+++ apps/inspector/src/components/credentials/index.ts
@@ -1,8 +1,4 @@
 export * from "./cas";
 export * from "./Credentials";
-export * from "./errors";
-export * from "./events";
 export * from "./get-content-type";
-export * from "./messaging";
-export * from "./types";
 export * from "./use-credentials";
