+++ apps/inspector/src/components/credentials/is-articlelike.ts
@@ -1,5 +1,5 @@
+import { ArticleLike } from "@originator-profile/extension-common";
 import { ContentAttestation } from "@originator-profile/model";
-import { ArticleLike } from "./types";
 
 export function isArticleLike<CA extends ContentAttestation>(
   sub: CA["credentialSubject"],
