+++ apps/inspector/src/components/credentials/use-credentials.ts
@@ -1,4 +1,11 @@
-import { getRegistry } from "@originator-profile/extension-common";
+import {
+  fetchTabCredentials,
+  fetchVerificationResult,
+  FrameIntegrityVerifier,
+  getRegistry,
+  type FramesVerifiedCas,
+  type SupportedVerifiedCas,
+} from "@originator-profile/extension-common";
 import type { OriginatorProfileSet } from "@originator-profile/model";
 import { verifyDocuments, type VerifiedOps } from "@originator-profile/verify";
 import { useParams } from "react-router";
@@ -6,12 +13,6 @@ import useSWRImmutable from "swr/immutable";
 import { toLegacyDocuments } from "../../utils/to-legacy-result";
 import { useSiteProfile } from "../siteProfile";
 import { deduplicateCas } from "./deduplicate-cas";
-import {
-  fetchTabCredentials,
-  fetchVerificationResult,
-  FrameIntegrityVerifier,
-} from "./messaging";
-import type { FramesVerifiedCas, SupportedVerifiedCas } from "./types";
 
 const CREDENTIALS_KEY = "credentials";
