+++ apps/inspector/src/components/link-verification/state.ts
@@ -1,5 +1,7 @@
-import { PersistentMap } from "@originator-profile/extension-common";
-import type { LinkVerificationResult } from "../credentials/types";
+import {
+  type LinkVerificationResult,
+  PersistentMap,
+} from "@originator-profile/extension-common";
 import type { PendingVerificationData, VerificationCacheData } from "./types";
 
 export const pendingOpIdVerification =
