+++ apps/inspector/src/components/overlay/FrameCaMarker.tsx
@@ -1,6 +1,6 @@
+import { SupportedVerifiedCa } from "@originator-profile/extension-common";
 import { WebMediaProfile } from "@originator-profile/model";
 import { twMerge } from "tailwind-merge";
-import { SupportedVerifiedCa } from "../credentials";
 import { CaCoordinate, FrameCoordinate } from "../frameCas/types";
 import { useFrameCaRects } from "../frameCas/use-frame-ca-rects";
 import { useLocatedCasCoordinate } from "../frameCas/use-located-cas-coordinate";
