+++ apps/inspector/src/components/siteProfile/index.ts
@@ -1,5 +1,3 @@
-export * from "./events";
-export * from "./messaging";
 export * from "./SiteProfile";
 export * from "./use-site-profile";
 export * from "./verify-website";
