+++ apps/inspector/src/components/siteProfile/types.ts
@@ -1,6 +1,4 @@
-import { type Serialized } from "@originator-profile/core";
 import { WebMediaProfile, WebsiteProfile } from "@originator-profile/model";
-import { FetchSiteProfileResult } from "@originator-profile/presentation";
 import { VerifiedSp } from "@originator-profile/verify";
 
 export type SiteProfileProps = {
@@ -9,5 +7,3 @@ export type SiteProfileProps = {
   wmp?: WebMediaProfile;
   wsp: WebsiteProfile;
 };
-
-export type SerializedSiteProfileResult = Serialized<FetchSiteProfileResult>;
