+++ apps/inspector/src/pages/Base.tsx
@@ -1,3 +1,8 @@
+import {
+  FetchCredentialsMessagingFailed,
+  FramesVerifiedCas,
+  SupportedVerifiedCas,
+} from "@originator-profile/extension-common";
 import { _ } from "@originator-profile/extension-common/ui";
 import {
   SiteProfileFetchFailed,
@@ -16,18 +21,13 @@ import {
 import flush from "just-flush";
 import { Navigate } from "react-router";
 import { useMount, useTitle } from "react-use";
-import Loading from "../components/Loading";
-import Unsupported from "../components/Unsupported";
-import {
-  FetchCredentialsMessagingFailed,
-  FramesVerifiedCas,
-  SupportedVerifiedCas,
-  useCredentials,
-} from "../components/credentials";
+import { useCredentials } from "../components/credentials";
 import { formatBuildModeTitle } from "../components/environment";
 import { useFrameCasLocationProvider } from "../components/frameCas";
+import Loading from "../components/Loading";
 import { overlayExtensionMessenger } from "../components/overlay/extension-events";
 import { useSiteProfile } from "../components/siteProfile";
+import Unsupported from "../components/Unsupported";
 import { buildPublUrl, routes } from "../utils/routes";
 
 function Redirect({
