+++ packages/extension-common/package.json
@@ -29,12 +29,14 @@
     "@iconify-react/fa6-solid": "1.0.2",
     "@originator-profile/core": "workspace:*",
     "@originator-profile/model": "workspace:*",
+    "@originator-profile/presentation": "workspace:*",
     "@originator-profile/securing-mechanism": "workspace:*",
     "@originator-profile/tsconfig": "workspace:*",
     "@originator-profile/verify": "workspace:*",
     "@types/chrome": "0.1.40",
     "@types/react": "19.2.14",
     "@types/react-dom": "19.2.3",
+    "@webext-core/messaging": "2.3.0",
     "dompurify": "3.4.13",
     "oxlint": "^1.73.0",
     "oxlint-tsgolint": "^0.24.0",
