+++ packages/extension-common/src/credentials/errors.ts
