+++ packages/extension-common/src/credentials/events.ts
