+++ packages/extension-common/src/credentials/index.ts
@@ -1,0 +1,4 @@
+export * from "./errors";
+export * from "./events";
+export * from "./messaging";
+export * from "./types";
