+++ packages/extension-common/src/credentials/messaging.ts
