+++ packages/extension-common/src/index.ts
@@ -1,4 +1,6 @@
+export * from "./credentials";
 export * from "./i18n";
+export * from "./site-profile";
 export * from "./utils/navigation-state";
 export * from "./utils/persistent-map";
 export * from "./utils/registry-ops";
