+++ packages/extension-common/src/site-profile/index.ts
@@ -1,0 +1,3 @@
+export * from "./events";
+export * from "./messaging";
+export * from "./types";
