+++ packages/extension-common/src/site-profile/types.ts
@@ -1,0 +1,4 @@
+import { type Serialized } from "@originator-profile/core";
+import { FetchSiteProfileResult } from "@originator-profile/presentation";
+
+export type SerializedSiteProfileResult = Serialized<FetchSiteProfileResult>;
